requirements.txt
tensorflow==2.9.0
numpy==1.21.6
Pillow==9.0.1
boto3==1.26.137
pymongo==4.3.3
python-multipart==0.0.5
serverless-python-requirements==5.4.0
pydantic==1.10.7
python-jose==3.3.0
cryptography==40.0.2